Creating HomeViewModel & HomeView refactor
