--- MyFitnessPal/Home/HomeView.swift
@@ -1,154 +1,0 @@
-//
-//  HomeView.swift
-//  MyFitnessPal
-//
-//  Created by Oualid AAMARRI on 05/02/2025.
-//
-
-import SwiftUI
-
-struct HomeView: View {
-    @State var calories: Int = 123
-    @State var active: Int = 52
-    @State var stand: Int = 8
-    
-    var mockActivities = [
-        Activity(id: 0, title: "Today steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .green, amount: "9812"),
-        Activity(id: 1, title: "Today steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .red, amount: "9812"),
-        Activity(id: 2, title: "Today steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .blue, amount: "9812"),
-        Activity(id: 3, title: "Today steps", subtitle: "Goal 12,000", image: "figure.walk", tintColor: .purple, amount: "9812")
-    ]
-    
-    var mockWorkouts = [
-        Workout(id: 0, title: "Running", image: "figure.run", tintcolor: .cyan, duration: "23 mins", date: "Aug 3", calories: "341 kcal"),
-        Workout(id: 1, title: "Running", image: "figure.run", tintcolor: .orange, duration: "23 mins", date: "Aug 3", calories: "341 kcal"),
-        Workout(id: 2, title: "Running", image: "figure.run", tintcolor: .yellow, duration: "23 mins", date: "Aug 3", calories: "341 kcal"),
-        Workout(id: 3, title: "Running", image: "figure.run", tintcolor: .brown, duration: "23 mins", date: "Aug 3", calories: "341 kcal")
-    ]
-    
-    var body: some View {
-        NavigationStack {
-            ScrollView(showsIndicators: false) {
-                VStack (alignment: .leading){
-                    Text("Welcome !")
-                        .font(.largeTitle)
-                        .padding()
-                    
-                    HStack {
-                        
-                        Spacer()
-                        
-                        VStack{
-                            VStack(alignment: .leading, spacing: 8) {
-                                Text("Calories")
-                                    .font(.callout)
-                                    .bold()
-                                    .foregroundStyle(.red)
-                                
-                                Text("123 Kcal")
-                                    .bold()
-                            }
-                            .padding(.bottom)
-                            
-                            VStack(alignment: .leading, spacing: 8) {
-                                Text("Active")
-                                    .font(.callout)
-                                    .bold()
-                                    .foregroundStyle(.green)
-                                
-                                Text("52 mins")
-                                    .bold()
-                            }
-                            .padding(.bottom)
-                            
-                            VStack(alignment: .leading, spacing: 8) {
-                                Text("Stand")
-                                    .font(.callout)
-                                    .bold()
-                                    .foregroundStyle(.blue)
-                                
-                                Text("8 hours")
-                                    .bold()
-                            }
-                            
-                        }
-                        
-                        Spacer()
-                        
-                        ZStack {
-                            
-                            ProgressCircleView(progress: $calories, goal: 600, color: .red)
-                            
-                            ProgressCircleView(progress: $active, goal: 60, color: .green)
-                                .padding(.all, 20)
-                            ProgressCircleView(progress: $stand, goal: 12, color: .blue)
-                                .padding(.all, 40)
-                        }
-                        
-                        .padding(.horizontal)
-                        
-                        Spacer()
-                        
-                    }
-                    .padding()
-                    
-                    HStack{
-                        Text("Fitness Activity")
-                            .font(.title2)
-                        
-                        Spacer()
-                        
-                        Button {
-                            print("show more")
-                        } label:{
-                            Text("Show more")
-                                .padding(.all, 10)
-                                .foregroundColor(.white)
-                                .background(.blue)
-                                .cornerRadius(20)
-                        }
-                    }
-                    .padding(.horizontal)
-                    
-                    LazyVGrid(columns: Array(repeating: GridItem(spacing: 20), count: 2)) {
-                        ForEach(mockActivities, id: \.id) { activity in
-                            ActivityCard(activity: activity)
-                        }
-                    }
-                    .padding(.horizontal)
-                    
-                    HStack{
-                        Text("Recent Workouts")
-                            .font(.title2)
-                        
-                        Spacer()
-                        
-                        NavigationLink {
-                            EmptyView()
-                        } label:{
-                            Text("Show more")
-                                .padding(.all, 10)
-                                .foregroundColor(.white)
-                                .background(.blue)
-                                .cornerRadius(20)
-                        }
-                    }
-                    .padding(.horizontal)
-                    .padding(.top)
-                    
-                    LazyVStack {
-                        ForEach(mockWorkouts, id: \.id) { workout in
-                            WorkoutCard(workout: workout)
-                        }
-                    }
-                    .padding(.bottom)
-                }
-                
-            }
-        }
-    }
-}
-
-#Preview {
-    HomeView()
-}
