+++ MyFitnessPal/Home/Models/Activity.swift
@@ -1,0 +1,18 @@
+//
+//  Activity.swift
+//  MyFitnessPal
+//
+//  Created by Oualid AAMARRI on 09/02/2025.
+//
+
+import SwiftUI
+
+
+struct Activity {
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tintColor: Color
+    let amount: String
+}
