+++ MyFitnessPal/Home/Models/Workout.swift
@@ -1,0 +1,19 @@
+//
+//  Workout.swift
+//  MyFitnessPal
+//
+//  Created by Oualid AAMARRI on 09/02/2025.
+//
+
+import SwiftUI
+
+
+struct Workout{
+    let id: Int
+    let title: String
+    let image: String
+    let tintcolor: Color
+    let duration: String
+    let date: String
+    let calories: String
+}
